Moved SanitizrNull.php out of Primitive dir
