+++ src/Sanitizr.php
@@ -4,12 +4,12 @@ namespace Nebalus\Sanitizr;
 
 use Nebalus\Sanitizr\Schema\AbstractSanitizrSchema;
 use Nebalus\Sanitizr\Schema\Primitive\SanitizrBoolean;
-use Nebalus\Sanitizr\Schema\Primitive\SanitizrNull;
 use Nebalus\Sanitizr\Schema\Primitive\SanitizrNumber;
 use Nebalus\Sanitizr\Schema\Primitive\SanitizrString;
 use Nebalus\Sanitizr\Schema\SanitizrArray;
 use Nebalus\Sanitizr\Schema\SanitizrBatch;
 use Nebalus\Sanitizr\Schema\SanitizrLiteral;
+use Nebalus\Sanitizr\Schema\SanitizrNull;
 use Nebalus\Sanitizr\Schema\SanitizrObject;
 
 class Sanitizr
